+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Added
+- Reading Topas4 formatted curves (from a directory of JSON files)
+- Setables now have the ability to have defaults, which are added to Notes if not otherwise specified.
+
 ### Fixed
 - Tune Test does not overwrite discrete tunes and runs instead of failing if they exist
 
@@ -15,6 +19,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 - Allow tunes to override setables from referenced arrangements (e.g. if idl refers to sig and idl also has setables that are defined in sig, use those defined in idl)
 
 ## [0.4.1]
+
+### Added
+- from_topas4  and associated test script, for import of Topas4 file format 
 
 ### Changed
 - setables are now fully optional